refactor(home): use SectionHeader for featured recipes
